Changing DB
DB has been changed to MySql, this will require you to create a DB at port 3306 with password and username root and root respectively. This can be done using MySql workbench and then creating a database schema named fusion. makemigrations and migrate after creating the database fusion for the app to work.