Database (arshad) (#9)
* Update README.md

* Updated Readme file

* Adding global database models.

* Correcting the line spacing and spellings

* Update README.md

* More informative readme

* Corrected using flake8 and isort

* Flke8 and isort compliant

* Adding EIS application

adding the employee information system app to the applications folder

* Changing DB

DB has been changed to MySql, this will require you to create a DB at port 3306 with password and username root and root respectively. This can be done using MySql workbench and then creating a database schema named fusion. makemigrations and migrate after creating the database fusion for the app to work.

* flake8 compliant

* Update models.py