So called Fix
The fix given here to filter by user and not working is against the comments written in HoldsDesignation Table models.py. It clearly says that eligibility must be matched through working column. We have done as a matter to speed up the review process, but technically then, it raises question on the comments written there,