Access to complaint module to unauthorised caretakers and supervisors
In the database, there were many users like faculties and acad admin who are assigned roles as caretakers and supervisors but are not authorised. Hence they show errors when trying to access the complaint module. So we fixed it by redirecting them to the complaint module as users so that they can at least file complaints.